requirements.txt
groq
pynput
python-dotenv>=1.0.0
sounddevice>=0.4.6
requests
opencc-python-reimplemented
PySide6>=6.7
psutil>=5.9
